Revert "Facilitator / Admin unified Course Room page incl. experimental icons, resource list styling"
