+++ client/src/Components/UI/ContentBox/contentBox.css
@@ -7,6 +7,8 @@
 }
 
 .Container {
+  min-width: 100%;
+  /* min-width: 540px; */
   position: relative;
   display: flex;
   padding: 5px;
@@ -18,6 +20,7 @@
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
+  width: 600px;
 }
 
 .SubContainer {
@@ -48,10 +51,7 @@
 
 .Container:hover {
   box-shadow: lightShadow;
-}
-
-.SubContainer:hover .Title {
-  font-weight: 600;
+  font-weight: 500;
 }
 
 .Title {
@@ -91,13 +91,6 @@
 .Expanded {
   margin-top: 5px;
   padding: 10px;
-}
-
-.DetailsTitle {
-  font-weight: 500;
-  margin-right: 5px;
-  /* width: 90px; */
-  display: inline-block;
 }
 
 .Icons {
