+++ client/src/Layout/Archive/Archive.js
@@ -94,94 +94,121 @@ const Archive = (props) => {
                 </div>
               </InfoBox>
             ) : null}
-            <div className={classes.TimePeriodLoadMoreContainer}>
-              <InfoBox
-                title="Time Period -- Last Updated More Than"
-                icon={<i className="fas fa-filter" />}
-              >
-                <Fragment>
-                  <div className={classes.FilterOpts}>
-                    <RadioBtn
-                      data-testid="all"
-                      check={() => toggleFilter('moreThan-all')}
-                      checked={dateRangePreset === 'all'}
-                      name="all"
-                    >
-                      All
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="day"
-                      check={() => toggleFilter('moreThan-afterDay')}
-                      checked={dateRangePreset === 'afterDay'}
-                      name="oneDay"
-                    >
-                      1 Day
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="one-week"
-                      check={() => toggleFilter('moreThan-afterWeek')}
-                      checked={dateRangePreset === 'afterWeek'}
-                      name="oneWeek"
-                    >
-                      1 Week
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="two-weeks"
-                      check={() => toggleFilter('moreThan-after2Weeks')}
-                      checked={dateRangePreset === 'after2Weeks'}
-                      name="two-weeks"
-                    >
-                      2 Weeks
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="one-month"
-                      check={() => toggleFilter('moreThan-afterMonth')}
-                      checked={dateRangePreset === 'afterMonth'}
-                      name="one-month"
-                    >
-                      1 Month
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="one-year"
-                      check={() => toggleFilter('moreThan-afterYear')}
-                      checked={dateRangePreset === 'afterYear'}
-                      name="one-year"
-                    >
-                      1 Year
-                    </RadioBtn>
-                    <RadioBtn
-                      data-testid="custom"
-                      check={() => toggleFilter('custom')}
-                      checked={dateRangePreset === 'custom'}
-                      name="custom"
-                    >
-                      Custom
-                    </RadioBtn>
-                  </div>
-                  {dateRangePreset === 'custom' ? (
-                    <div className={classes.CustomDateInputs}>
-                      <div className={classes.CustomFromDate}>
-                        <h3>From Date</h3>
-                        <DatePicker
-                          selected={customFromDate}
-                          onChange={setFromDate}
-                          popperPlacement="bottom"
-                          shouldCloseOnSelect
-                        />
-                      </div>
-                      <div className={classes.CustomToDate}>
-                        <h3>To Date</h3>
-                        <DatePicker
-                          selected={customToDate}
-                          onChange={setToDate}
-                          popperPlacement="bottom"
-                          shouldCloseOnSelect
-                        />
-                      </div>
+            <InfoBox
+              title="Time Period -- Last Updated More Than"
+              icon={<i className="fas fa-filter" />}
+            >
+              <Fragment>
+                <div className={classes.FilterOpts}>
+                  <RadioBtn
+                    data-testid="all"
+                    check={() => toggleFilter('moreThan-all')}
+                    checked={dateRangePreset === 'all'}
+                    name="all"
+                  >
+                    All
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="day"
+                    check={() => toggleFilter('moreThan-afterDay')}
+                    checked={dateRangePreset === 'afterDay'}
+                    name="oneDay"
+                  >
+                    1 Day
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="one-week"
+                    check={() => toggleFilter('moreThan-afterWeek')}
+                    checked={dateRangePreset === 'afterWeek'}
+                    name="oneWeek"
+                  >
+                    1 Week
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="two-weeks"
+                    check={() => toggleFilter('moreThan-after2Weeks')}
+                    checked={dateRangePreset === 'after2Weeks'}
+                    name="two-weeks"
+                  >
+                    2 Weeks
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="one-month"
+                    check={() => toggleFilter('moreThan-afterMonth')}
+                    checked={dateRangePreset === 'afterMonth'}
+                    name="one-month"
+                  >
+                    1 Month
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="one-year"
+                    check={() => toggleFilter('moreThan-afterYear')}
+                    checked={dateRangePreset === 'afterYear'}
+                    name="one-year"
+                  >
+                    1 Year
+                  </RadioBtn>
+                  <RadioBtn
+                    data-testid="custom"
+                    check={() => toggleFilter('custom')}
+                    checked={dateRangePreset === 'custom'}
+                    name="custom"
+                  >
+                    Custom
+                  </RadioBtn>
+                </div>
+                {dateRangePreset === 'custom' ? (
+                  <div className={classes.CustomDateInputs}>
+                    <div className={classes.CustomFromDate}>
+                      <h3>From Date</h3>
+                      <DatePicker
+                        selected={customFromDate}
+                        onChange={setFromDate}
+                        popperPlacement="bottom"
+                        shouldCloseOnSelect
+                      />
                     </div>
-                  ) : null}
-                </Fragment>
-              </InfoBox>
+                    <div className={classes.CustomToDate}>
+                      <h3>To Date</h3>
+                      <DatePicker
+                        selected={customToDate}
+                        onChange={setToDate}
+                        popperPlacement="bottom"
+                        shouldCloseOnSelect
+                      />
+                    </div>
+                  </div>
+                ) : null}
+              </Fragment>
+            </InfoBox>
+          </div>
+        </div>
+        <div
+          className={classes.List}
+          // style={{
+          //   marginTop: header.current
+          //     ? header.current.getBoundingClientRect().height
+          //     : 260,
+          // }}
+        >
+          {/* Check to see if visibleResources is undef, which is the loading state */}
+          {loading ? (
+            <div className={classes.Pending}>
+              Loading
+              <span className={classes.dot1}>.</span>
+              <span className={classes.dot2}>.</span>
+              <span className={classes.dot3}>.</span>
+            </div>
+          ) : (
+            <div style={{ display: 'flex', flexDirection: 'column' }}>
+              <SelectableBoxList
+                list={visibleResources}
+                resource={resource}
+                listType="public"
+                selectable
+                icons={icons}
+                selectActions={selectActions}
+              />
               <div className={classes.LoadMore}>
                 <Button
                   m={20}
@@ -196,35 +223,8 @@ const Archive = (props) => {
                 </Button>
               </div>
             </div>
-          </div>
+          )}
         </div>
-        {/* <div
-          className={classes.List}
-          // style={{
-          //   marginTop: header.current
-          //     ? header.current.getBoundingClientRect().height
-          //     : 260,
-          // }}
-        > */}
-        {/* Check to see if visibleResources is undef, which is the loading state */}
-        {loading ? (
-          <div className={classes.Pending}>
-            Loading
-            <span className={classes.dot1}>.</span>
-            <span className={classes.dot2}>.</span>
-            <span className={classes.dot3}>.</span>
-          </div>
-        ) : (
-          <SelectableBoxList
-            list={visibleResources}
-            resource={resource}
-            listType="public"
-            selectable
-            icons={icons}
-            selectActions={selectActions}
-          />
-        )}
-        {/* </div> */}
       </div>
     </React.Fragment>
   );
