+++ client/src/Layout/BoxList/boxList.css
@@ -11,8 +11,10 @@
 }
 
 .ContentBox {
-  max-width: 1000px;
-  min-width: 573px;
+  min-width: 100%;
+  width: 600px;
+  /* min-width: 497px; */
+  /* max-width: 33vw; */
   margin: 2px;
   transition: 0.2s;
 }
