+++ client/src/Layout/Community/Community.js
@@ -103,32 +103,22 @@ class Community extends Component {
               </div>
             </InfoBox>
             {resource !== 'courses' ? (
-              <div className={classes.RoomTypeLoadMoreContainer}>
-                <InfoBox
-                  title="Room Type"
-                  icon={<i className="fas fa-filter" />}
-                >
-                  <div className={classes.FilterOpts}>
-                    <RadioBtn
-                      data-testid="all-roomType-filter"
-                      check={() => toggleFilter('all-roomType')}
-                      checked={filters.roomType === 'all'}
-                      name="All-roomType"
-                    >
-                      All
-                    </RadioBtn>
-                    <TabTypes.RadioButtons
-                      onClick={toggleFilter}
-                      checked={filters.roomType}
-                    />
-                  </div>
-                </InfoBox>
-                <div className={classes.LoadMore}>
-                  <Button m={20} disabled={!moreAvailable} click={setSkip}>
-                    load more results
-                  </Button>
+              <InfoBox title="Room Type" icon={<i className="fas fa-filter" />}>
+                <div className={classes.FilterOpts}>
+                  <RadioBtn
+                    data-testid="all-roomType-filter"
+                    check={() => toggleFilter('all-roomType')}
+                    checked={filters.roomType === 'all'}
+                    name="All-roomType"
+                  >
+                    All
+                  </RadioBtn>
+                  <TabTypes.RadioButtons
+                    onClick={toggleFilter}
+                    checked={filters.roomType}
+                  />
                 </div>
-              </div>
+              </InfoBox>
             ) : null}
           </div>
         </div>
@@ -157,6 +147,11 @@ class Community extends Component {
                 linkSuffix={linkSuffix}
                 listType="public"
               />
+              <div className={classes.LoadMore}>
+                <Button m={20} disabled={!moreAvailable} click={setSkip}>
+                  load more results
+                </Button>
+              </div>
             </Fragment>
           )}
         </div>
