+++ client/src/Layout/Dashboard/MainContent/ResourceList.js
@@ -1,6 +1,7 @@
-import React, { Fragment, useEffect, useState } from 'react';
+import React, { Fragment, useEffect, useState, useRef } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
+import Select from 'react-select';
 import PropTypes from 'prop-types';
 import {
   useSortableData,
@@ -9,7 +10,7 @@ import {
   useUIState,
 } from 'utils';
 import SelectableBoxList from 'Layout/SelectableBoxList/SelectableBoxList';
-import { Button, BigModal, Modal, SortUI, ToolTip } from 'Components';
+import { Button, Modal, BigModal, Search, ToolTip } from 'Components';
 import { RoomPreview } from 'Containers';
 import { updateRoom, archiveRooms } from 'store/actions';
 import { STATUS } from 'constants.js';
@@ -337,71 +338,67 @@ const ResourceList = ({
         <div className={classes.Controls}>{create}</div>
         {fList.length > 0 || pList.length > 0 ? (
           <div className={classes.Row}>
-            {fList.length > 0 ? (
-              <div className={classes.Col}>
-                <h2 className={classes.ResourceHeader}>
-                  {displayResource} I Manage: {fList.length}
-                </h2>
-                {fList.length >= 1 && setResourceState && (
-                  <SortUI
-                    keys={keys}
-                    sortFn={facilitatorRequestSort}
-                    sortConfig={facilitatorSortConfig}
-                  />
-                )}
+            <div className={classes.Col}>
+              <h2 className={classes.ResourceHeader}>
+                {displayResource} I Manage: {fList.length}
+              </h2>
+              {fList.length >= 1 && setResourceState && (
+                <SortUI
+                  keys={keys}
+                  sortFn={facilitatorRequestSort}
+                  sortConfig={facilitatorSortConfig}
+                />
+              )}
 
-                {selectableBoxList &&
-                resource !== 'activities' &&
-                resource !== 'courses' ? (
-                  <SelectableBoxList
-                    list={facilitatorItems}
-                    resource={resource}
-                    listType="private"
-                    linkPath={linkPath}
-                    linkSuffix={linkSuffix}
-                    notifications={notifications}
-                    parentResourec={parentResource}
-                    icons={customIcons}
-                    selectActions={selectActions}
-                  />
-                ) : (
-                  <BoxList
-                    list={facilitatorItems}
-                    resource={resource}
-                    listType="private"
-                    linkPath={linkPath}
-                    linkSuffix={linkSuffix}
-                    notifications={notifications}
-                    parentResourec={parentResource}
-                  />
-                )}
-              </div>
-            ) : null}
-            {pList.length > 0 ? (
-              <div className={classes.Col}>
-                <h2 className={classes.ResourceHeader}>
-                  {displayResource} I&#39;m a member of: {pList.length}
-                </h2>
-                {pList.length >= 1 && setResourceState && (
-                  <SortUI
-                    keys={keys}
-                    sortFn={participantRequestSort}
-                    sortConfig={participantSortConfig}
-                  />
-                )}
-                <BoxList
-                  list={participantItems}
+              {selectableBoxList &&
+              resource !== 'activities' &&
+              resource !== 'courses' ? (
+                <SelectableBoxList
+                  list={facilitatorItems}
+                  resource={resource}
+                  listType="private"
                   linkPath={linkPath}
                   linkSuffix={linkSuffix}
                   notifications={notifications}
+                  parentResourec={parentResource}
+                  icons={customIcons}
+                  selectActions={selectActions}
+                />
+              ) : (
+                <BoxList
+                  list={facilitatorItems}
                   resource={resource}
                   listType="private"
+                  linkPath={linkPath}
+                  linkSuffix={linkSuffix}
+                  notifications={notifications}
                   parentResourec={parentResource}
-                  icons={customIconsBoxList}
-                  // draggable
                 />
-              </div>
-            ) : null}
+              )}
+            </div>
+            <div className={classes.Col}>
+              <h2 className={classes.ResourceHeader}>
+                {displayResource} I&#39;m a member of: {pList.length}
+              </h2>
+              {pList.length >= 1 && setResourceState && (
+                <SortUI
+                  keys={keys}
+                  sortFn={participantRequestSort}
+                  sortConfig={participantSortConfig}
+                />
+              )}
+              <BoxList
+                list={participantItems}
+                linkPath={linkPath}
+                linkSuffix={linkSuffix}
+                notifications={notifications}
+                resource={resource}
+                listType="private"
+                parentResourec={parentResource}
+                icons={customIconsBoxList}
+                // draggable
+              />
+            </div>
           </div>
         ) : (
           <Fragment>
@@ -465,3 +462,192 @@ ResourceList.defaultProps = {
 };
 
 export default ResourceList;
+
+// Currently, SortUI is only used within ResourceList. Hence, why it's here.
+const SortUI = ({ keys, sortFn, sortConfig }) => {
+  const upArrow = <i className="fas fa-solid fa-arrow-up" />;
+  const downArrow = <i className="fas fa-solid fa-arrow-down" />;
+  const timeFrameOptions = [
+    { label: 'All', value: timeFrames.ALL },
+    { label: 'Last Day', value: timeFrames.LASTDAY },
+    { label: 'Last Week', value: timeFrames.LASTWEEK },
+    { label: 'Last Two Weeks', value: timeFrames.LAST2WEEKS },
+    { label: 'Last Month', value: timeFrames.LASTMONTH },
+    { label: 'Last Year', value: timeFrames.LASTYEAR },
+    { label: 'More than a Day', value: timeFrames.AFTERDAY },
+    { label: 'More than a Week', value: timeFrames.AFTERWEEK },
+    { label: 'More than Two Weeks', value: timeFrames.AFTER2WEEKS },
+    { label: 'More than a Month', value: timeFrames.AFTERMONTH },
+    { label: 'More than a Year', value: timeFrames.AFTERYEAR },
+  ];
+
+  const previousSearch = useRef({
+    criteria: sortConfig.criteria || '',
+    filter: {
+      ...sortConfig.filter,
+      filterFcn:
+        sortConfig.criteria && sortConfig.criteria !== ''
+          ? (item) =>
+              item.name &&
+              item.name
+                .toLowerCase()
+                .indexOf(sortConfig.criteria.toLowerCase()) > -1
+          : null,
+    },
+  });
+
+  React.useEffect(() => {
+    if (
+      previousSearch.current.criteria &&
+      previousSearch.current.criteria !== ''
+    ) {
+      previousSearch.current.filter.filterFcn = (item) =>
+        item.name &&
+        item.name
+          .toLowerCase()
+          .indexOf(previousSearch.current.criteria.toLowerCase()) > -1;
+    } else {
+      previousSearch.current.filter.filterFcn = null;
+    }
+  }, [previousSearch.current.criteria]);
+
+  const optionForValue = (value) => {
+    return timeFrameOptions.find((opt) => opt.value === value);
+  };
+
+  const keyName = (defaultName) => {
+    const matchingKey = keys.find((key) => key.property === sortConfig.key);
+    return matchingKey ? matchingKey.name : defaultName;
+  };
+
+  const labelSuffix = Math.ceil(Math.random() * 1000);
+
+  const search = (criteria) => {
+    if (criteria !== '' && previousSearch.current.criteria === '') {
+      previousSearch.current = {
+        criteria,
+        filter: sortConfig.filter,
+      };
+      sortFn({
+        criteria,
+        filter: {
+          ...sortConfig.filter,
+          timeframe: timeFrames.ALL,
+          filterFcn: (item) =>
+            item.name &&
+            item.name.toLowerCase().indexOf(criteria.toLowerCase()) > -1,
+        },
+      });
+    } else if (criteria !== '' && previousSearch.current.criteria !== '') {
+      sortFn({
+        criteria,
+        filter: {
+          ...sortConfig.filter,
+          filterFcn: (item) =>
+            item.name &&
+            item.name.toLowerCase().indexOf(criteria.toLowerCase()) > -1,
+        },
+      });
+    } else if (criteria === '' && previousSearch.current.criteria !== '') {
+      sortFn({
+        criteria,
+        filter: { ...previousSearch.current.filter, filterFcn: null },
+      });
+      previousSearch.current.criteria = '';
+    }
+  };
+
+  return (
+    <div className={classes.SortUIContainer}>
+      <div className={classes.SortSelection}>
+        <label htmlFor={`sortUI-${labelSuffix}`} className={classes.Label}>
+          Sort by:
+          <Select
+            className={classes.Select}
+            inputId={`sortUI-${labelSuffix}`}
+            placeholder="Select..."
+            onChange={(selectedOption) => {
+              sortFn({
+                key: selectedOption.value,
+                direction: sortConfig.direction,
+              });
+            }}
+            value={{
+              // eslint-disable-next-line react/prop-types
+              label: keyName(keys[0].name),
+              // eslint-disable-next-line react/prop-types
+              value: sortConfig.key || keys[0].property,
+            }}
+            options={keys.map((key) => ({
+              value: key.property,
+              label: key.name,
+            }))}
+            isSearchable={false}
+          />{' '}
+        </label>
+        <span
+          style={{ padding: '0 5px' }}
+          onClick={() => sortFn({ key: sortConfig.key })}
+          onKeyDown={() => sortFn({ key: sortConfig.key })}
+          role="button"
+          tabIndex={-1}
+        >
+          {sortConfig.direction === 'descending' ? downArrow : upArrow}{' '}
+        </span>
+      </div>
+      <div className={classes.FilterSelection}>
+        <label
+          htmlFor={`filterUI-${labelSuffix}`}
+          className={classes.Label}
+          data-testid={`filterUI-${labelSuffix}`}
+        >
+          Updated:
+          <Select
+            placeholder="Timeframe"
+            className={classes.Select}
+            inputId={`filterUI-${labelSuffix}`}
+            onChange={(selectedOption) => {
+              sortFn({
+                filter: {
+                  ...sortConfig.filter,
+                  timeframe: selectedOption.value,
+                },
+              });
+            }}
+            value={optionForValue(sortConfig.filter.timeframe)}
+            options={timeFrameOptions}
+            isSearchable={false}
+          />{' '}
+        </label>
+      </div>
+      <div className={classes.Search}>
+        <Search
+          isControlled
+          value={sortConfig.criteria || ''}
+          _search={search}
+          data-testid="search"
+        />
+      </div>
+    </div>
+  );
+};
+
+SortUI.propTypes = {
+  keys: PropTypes.arrayOf(
+    PropTypes.shape({ name: PropTypes.string, property: PropTypes.string })
+  ).isRequired,
+  sortFn: PropTypes.func.isRequired,
+  sortConfig: PropTypes.shape({
+    criteria: PropTypes.string,
+    key: PropTypes.string,
+    direction: PropTypes.string,
+    filter: PropTypes.shape({
+      key: PropTypes.string,
+      timeframe: PropTypes.string,
+      filterFcn: PropTypes.func,
+    }),
+  }),
+};
+SortUI.defaultProps = {
+  sortConfig: {},
+};
