+++ client/src/Layout/SelectableBoxList/SelectableBoxList.css
@@ -13,7 +13,6 @@
   margin: 10px 0;
   width: 100%;
   max-width: 1100px;
-  /* max-width: 575px; */
   color: #444;
   background: #f8f8f8;
   box-shadow: 10px 10px 10px 2px #f8f8f8, -10px 0 10px 2px #f8f8f8;
@@ -31,12 +30,8 @@
 }
 
 .ContentBox {
-  max-width: 850px;
+  min-width: 100%;
+  width: 580px;
   margin: 2px 2px 2px 0;
   transition: 0.2s;
-}
-
-.SelectActions {
-  display: flex;
-
 }
