+++ client/src/Layout/SelectableBoxList/SelectableBoxList.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { Checkbox } from 'Components';
 import SelectableContentBox from 'Components/UI/ContentBox/SelectableContentBox';
@@ -14,49 +14,10 @@ const SelectableBoxList = (props) => {
     linkSuffix,
     selectActions, // array of actions (Buttons) i.e. Restore, Delete
     icons,
-    customStyle,
   } = props;
 
   const [selectedIds, setSelectedIds] = useState([]);
   const [selectAllChecked, setSelectAllChecked] = useState(false);
-  const [formattedCustomStyles, setFormattedCustomStyles] = useState({});
-
-  useEffect(() => {
-    if (!customStyle || !Object.keys(customStyle).length) return;
-
-    const getCustomElementStyle = () => {
-      // this fn returns a formatted copy of customStyle
-      // which will be used in the render if needed
-
-      // customStyle is a prop object
-      // { classnameToStyle: cssObject }
-      const formattedCustomStyle = {};
-      const classNames = {
-        container: 'Container',
-        header: 'Header',
-        selectactions: 'SelectActions',
-        contentbox: 'ContentBox',
-        title: 'Title',
-      };
-
-      const customStyleKeys = Object.keys(customStyle);
-      customStyleKeys.forEach((elementToStyle) => {
-        // if (!classNames.includes(elementToStyle.toLowerCase)) return null;
-        formattedCustomStyle[elementToStyle] = {
-          ...formattedCustomStyle[elementToStyle.toLocaleLowerCase()],
-          ...customStyle[elementToStyle],
-        };
-      });
-      return formattedCustomStyle;
-    };
-
-    setFormattedCustomStyles(getCustomElementStyle(customStyle));
-  }, [customStyle]);
-
-  useEffect(() => {
-    // deselect all if length of items to display changes
-    handleDeselectAll();
-  }, [list.length]);
 
   const handleSelectAll = (event) => {
     const { checked } = event.target;
@@ -81,14 +42,6 @@ const SelectableBoxList = (props) => {
       setSelectedIds((prevState) => [...prevState.filter((el) => id !== el)]);
       setSelectAllChecked(false);
     }
-  };
-
-  const handleDeselectAll = () => {
-    // this function is passed up to the parent component
-    // and can be used to deselect all checkboxes once
-    // certain selectAction onClicks are triggered
-    setSelectAllChecked(false);
-    setSelectedIds([]);
   };
 
   const timeDiff = (ts) => {
@@ -121,137 +74,102 @@ const SelectableBoxList = (props) => {
   }
 
   return (
-    <div className={classes.Container} style={formattedCustomStyles.container}>
-      <div className={classes.Header} style={formattedCustomStyles.header}>
+    <div className={classes.Container}>
+      <div className={classes.Header}>
         <Checkbox
           change={handleSelectAll}
           checked={selectAllChecked}
           dataId="select-all"
-          style={formattedCustomStyles.checkbox}
         >
           Select All
         </Checkbox>
-        <div
-          className={classes.SelectActions}
-          style={formattedCustomStyles.selectactions}
-        >
-          {selectActions.map((selectAction) => (
-            <div
-              onClick={(e) => {
-                selectAction.onClick(e, selectedIds, handleDeselectAll);
-              }}
-              onKeyDown={(e) => {
-                selectAction.onClick(e, selectedIds, handleDeselectAll);
-              }}
-              role="button"
-              tabIndex={-1}
-              // title={selectAction.title}
-              key={`selectAction-${selectAction.title}`}
-              style={{
-                margin: '0 1rem',
-                ...formattedCustomStyles[selectAction.title],
-              }}
-              data-testid={`${selectAction.title}-icon`}
-              // title={`${selectAction.title}-icon`}
-            >
-              {selectAction.generateIcon
-                ? selectAction.generateIcon(selectedIds)
-                : selectAction.icon}
-            </div>
-          ))}
-        </div>
+        {selectActions.map((selectAction) => (
+          <div
+            onClick={(e) => selectAction.onClick(e, selectedIds)}
+            onKeyDown={(e) => selectAction.onClick(e, selectedIds)}
+            role="button"
+            tabIndex={-1}
+            // title={selectAction.title}
+            key={`selectAction-${selectAction.title}`}
+            style={{ margin: '0 1rem' }}
+            data-testid={`${selectAction.title}-icon`}
+            title={`${selectAction.title}-icon`}
+          >
+            {selectAction.icon}
+          </div>
+        ))}
       </div>
+      {list.map((item) => {
+        if (item) {
+          const details = {
+            description: item.description,
+            createdAt: item.createdAt
+              ? item.createdAt.split('T')[0].toLocaleString()
+              : '',
+            dueDate: item.dueDate,
+            facilitators: item.members
+              ? item.members
+                  .filter((member) => member.role === 'facilitator')
+                  .map(
+                    (member, x, arr) =>
+                      `${
+                        member.user
+                          ? member.user.username
+                          : 'Username not found'
+                      }${x < arr.length - 1 ? ', ' : ''}`
+                  )
+              : [],
+            sinceUpdated: timeDiff(item.updatedAt),
+          };
 
-      <div className={classes.BoxList}>
-        {list.map((item) => {
-          if (item) {
-            const details = {
-              description: item.description,
-              createdAt: item.createdAt
-                ? item.createdAt.split('T')[0].toLocaleString()
-                : '',
-              dueDate: item.dueDate,
-              facilitators: item.members
-                ? item.members
-                    .filter((member) => member.role === 'facilitator')
-                    .map(
-                      (member, x, arr) =>
-                        `${
-                          member.user
-                            ? member.user.username
-                            : 'Username not found'
-                        }${x < arr.length - 1 ? ', ' : ''}`
-                    )
-                : [],
-              participants: item.members
-                ? item.members
-                    .filter((member) => member.role === 'participant')
-                    .map(
-                      (member, x, arr) =>
-                        `${
-                          member.user
-                            ? member.user.username
-                            : 'Username not found'
-                        }${x < arr.length - 1 ? ', ' : ''}`
-                    )
-                : [],
-              sinceUpdated: timeDiff(item.updatedAt),
-            };
-
-            let notificationCount = 0;
-            if (listType === 'private') {
-              if (notifications.length > 0) {
-                notifications.forEach((ntf) => {
-                  if (
-                    ntf.resourceId === item._id ||
-                    ntf.parentResource === item._id
-                  ) {
-                    notificationCount += 1;
-                  }
-                });
-              }
-              details.entryCode = item.entryCode;
-            } else if (item.creator) {
-              details.creator = item.creator.username;
+          let notificationCount = 0;
+          if (listType === 'private') {
+            if (notifications.length > 0) {
+              notifications.forEach((ntf) => {
+                if (
+                  ntf.resourceId === item._id ||
+                  ntf.parentResource === item._id
+                ) {
+                  notificationCount += 1;
+                }
+              });
             }
-            return (
-              <div
-                className={classes.ContentBox}
-                key={item._id}
-                style={formattedCustomStyles.contentbox}
-              >
-                <SelectableContentBox
-                  title={item.name}
-                  link={
-                    linkPath !== null
-                      ? `${linkPath}${item._id}${linkSuffix}`
-                      : null
-                  }
-                  key={item._id}
-                  id={item._id}
-                  isChecked={selectedIds.includes(item._id)}
-                  onSelect={handleSelectOne}
-                  notifications={notificationCount}
-                  roomType={
-                    item && item.tabs
-                      ? item.tabs.map((tab) => tab.tabType)
-                      : item.tabTypes
-                  }
-                  locked={item.privacySetting === 'private'}
-                  details={details}
-                  listType={listType}
-                  customIcons={icons}
-                  resource={resource}
-                  customStyle={item.customStyle}
-                >
-                  {item.description}
-                </SelectableContentBox>
-              </div>
-            );
+            details.entryCode = item.entryCode;
+          } else if (item.creator) {
+            details.creator = item.creator.username;
           }
-          return null;
-        })}
-      </div>
+          return (
+            <div className={classes.ContentBox} key={item._id}>
+              <SelectableContentBox
+                title={item.name}
+                link={
+                  linkPath !== null
+                    ? `${linkPath}${item._id}${linkSuffix}`
+                    : null
+                }
+                key={item._id}
+                id={item._id}
+                isChecked={selectedIds.includes(item._id)}
+                onSelect={handleSelectOne}
+                notifications={notificationCount}
+                roomType={
+                  item && item.tabs
+                    ? item.tabs.map((tab) => tab.tabType)
+                    : item.tabTypes
+                }
+                locked={item.privacySetting === 'private'} // @TODO Should it appear locked if the user has access ? I can see reasons for both
+                details={details}
+                listType={listType}
+                customIcons={icons}
+                resource={resource}
+              >
+                {item.description}
+              </SelectableContentBox>
+            </div>
+          );
+        }
+        return null;
+      })}
     </div>
   );
 };
@@ -266,7 +184,6 @@ SelectableBoxList.propTypes = {
   selectedIds: PropTypes.arrayOf(PropTypes.string),
   selectActions: PropTypes.arrayOf(PropTypes.shape({})),
   icons: PropTypes.arrayOf(PropTypes.shape({})),
-  customStyle: PropTypes.shape({}),
 };
 
 SelectableBoxList.defaultProps = {
@@ -276,7 +193,6 @@ SelectableBoxList.defaultProps = {
   icons: null,
   linkPath: null,
   linkSuffix: null,
-  customStyle: null,
 };
 
 export default SelectableBoxList;
